Move FromJson test methods down a layer
