+++ src/Microsoft.VisualStudio.ProjectSystem.Managed.UnitTests/Mocks/ProjectConfigurationFactory.cs
@@ -40,6 +40,12 @@ namespace Microsoft.VisualStudio.ProjectSystem
             return configurationsBuilder.ToImmutable();
         }
 
+        public static ProjectConfiguration FromJson(string jsonString)
+        {
+            var model = new ProjectConfigurationModel();
+            return model.FromJson(jsonString);
+        }
+
 
         private static string GetDimensionName(int ordinal)
         {
@@ -56,6 +62,34 @@ namespace Microsoft.VisualStudio.ProjectSystem
 
                 default:
                     throw new InvalidOperationException();
+            }
+        }
+    }
+
+    internal class ProjectConfigurationModel : JsonModel<ProjectConfiguration>
+    {
+        public IImmutableDictionary<string, string> Dimensions { get; set; }
+
+        public string Name { get; set; }
+
+        public override ProjectConfiguration ToActualModel()
+        {
+            return new ActualModel
+            {
+                Dimensions = Dimensions,
+                Name = Name
+            };
+        }
+
+        private class ActualModel : ProjectConfiguration
+        {
+            public IImmutableDictionary<string, string> Dimensions { get; set; }
+
+            public string Name { get; set; }
+
+            public bool Equals(ProjectConfiguration other)
+            {
+                throw new NotImplementedException();
             }
         }
     }
